+++ packages/rspeedy/core/src/config/defaults.ts
@@ -10,6 +10,10 @@ import type { Filename } from './output/filename.js'
 import type { Config } from './index.js'
 
 export function applyDefaultRspeedyConfig(config: Config): Config {
+  // config.performance?.chunkSplit?.strategy has been explicitly set to a value other than 'all-in-one'
+  const enableChunkSplitting = config.performance?.chunkSplit?.strategy
+    && config.performance?.chunkSplit?.strategy !== 'all-in-one'
+
   return mergeRsbuildConfig({
     mode: ((): RsbuildMode => {
       if (config.mode) {
@@ -26,8 +30,8 @@ export function applyDefaultRspeedyConfig(config: Config): Config {
       // from the `output.filename.bundle` field.
       filename: getFilename(config.output?.filename),
 
-      // inlineScripts should be enabled by default
-      inlineScripts: true,
+      // inlineScripts defaults to false when chunk splitting is enabled, true otherwise
+      inlineScripts: !enableChunkSplitting,
     },
 
     performance: {