+++ packages/rspeedy/plugin-react-alias/src/index.ts
@@ -27,13 +27,6 @@ export function pluginReactAlias(options: Options): RsbuildPlugin {
   return {
     name: 'lynx:react-alias',
     setup(api) {
-      const hasAlias = api.useExposed<boolean>(S_PLUGIN_REACT_ALIAS)
-      if (hasAlias) {
-        // We make sure that only make aliased once
-        return
-      }
-      api.expose(S_PLUGIN_REACT_ALIAS, true)
-
       const require = createRequire(import.meta.url)
 
       const reactLynxPkg = require.resolve('@lynx-js/react/package.json', {
@@ -56,7 +49,14 @@ export function pluginReactAlias(options: Options): RsbuildPlugin {
         })
       })
 
-      api.modifyBundlerChain(async (chain, { isProd }) => {
+      api.modifyBundlerChain(async (chain, { isProd, environment }) => {
+        if (Object.hasOwn(environment, S_PLUGIN_REACT_ALIAS)) {
+          // This environment has already been processed
+          return
+        }
+        Object.defineProperty(environment, S_PLUGIN_REACT_ALIAS, {
+          value: true,
+        })
         const [
           jsxRuntimeBackground,
           jsxRuntimeMainThread,